+++ src/main/java/com/atlauncher/utils/CurseApi.java
@@ -37,8 +37,10 @@ public class CurseApi {
     public static List<CurseMod> searchMods(String gameVersion, String query, int page, int categoryId) {
         try {
             String url = String.format(
-                    "%s/addon/search?gameId=432&gameVersion=%s&categoryId=%d&sectionId=6&searchFilter=%s&sort=Popularity&sortDescending=true&pageSize=%d&index=%d",
+                    "%s/addon/search?gameId=432&gameVersion=%s&categoryId=%d&sectionId=%s&searchFilter=%s&sort=Popularity&sortDescending=true&pageSize=%d&index=%d",
                     Constants.CURSE_API_URL, gameVersion, categoryId,
+                    URLEncoder.encode(String.format("%d,%d", Constants.CURSE_MODS_SECTION_ID,
+                            Constants.CURSE_RESOURCE_PACKS_SECTION_ID), StandardCharsets.UTF_8.name()),
                     URLEncoder.encode(query, StandardCharsets.UTF_8.name()), Constants.CURSE_PAGINATION_SIZE,
                     page * Constants.CURSE_PAGINATION_SIZE);