+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 -   Create profile.json when installing instances which follows same pattern as the Minecraft one
 -   Rework the way downloads are handled
 -   Remove the option to download from multiple ATLauncher servers
+-   Add support for adding resource packs from Curse